。。bug

+++ src/main/java/com/stdwork_management/controller/ClientController.java
@@ -11,6 +11,7 @@ import io.swagger.annotations.ApiParam;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang.StringUtils;
 import org.apache.tomcat.util.http.fileupload.FileUtils;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.web.bind.annotation.*;
@@ -47,6 +48,13 @@ public class ClientController {
 
     @Autowired
     private String workPath;
+
+    private static final String datePattern1 = "-(([0-9]{3}[1-9]|[0-9]{2}[1-9][0-9]{1}|[0-9]{1}[1-9][0-9]{2}|[1-9][0-9]{3})(((0[13578]|" +
+            "1[02])(0[1-9]|[12][0-9]|3[01]))|((0[469]|11)(0[1-9]|[12][0-9]|30))|(02-(0[1-9]|[1][0-9]|2[0-8])))" +
+            ")|((([0-9]{2})(0[48]|[2468][048]|[13579][26])|((0[48]|[2468][048]|[3579][26])00))0229)(([0-9]{3}[" +
+            "1-9]|[0-9]{2}[1-9][0-9]{1}|[0-9]{1}[1-9][0-9]{2}|[1-9][0-9]{3})(((0[13578]|1[02])(0[1-9]|[12][0-9" +
+            "]|3[01]))|((0[469]|11)(0[1-9]|[12][0-9]|30))|(02-(0[1-9]|[1][0-9]|2[0-8]))))|((([0-9]{2})(0[48]|[2" +
+            "468][048]|[13579][26])|((0[48]|[2468][048]|[3579][26])00))0229)$";
 
     @Autowired
     private LocalFileSysMapper localFileSysMapper;
@@ -135,15 +143,24 @@ public class ClientController {
         Condition condition = new Condition(LocalFileSysPO.class);
         Example.Criteria criteria = condition.or();
         localFileSysPOList.forEach(po -> criteria.orLike("path", po.getPath().substring(0, po.getPath().lastIndexOf("-") + 1) + "%"));
-        Optional.ofNullable(localFileSysMapper.selectByCondition(condition)).orElse(new ArrayList<>())
-                .forEach( po -> {
-                    try {
-                        FileUtils.forceDelete(new File(po.getPath()));
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                });
-        localFileSysMapper.deleteByCondition(condition);
+        List<LocalFileSysPO> collect = Optional.ofNullable(localFileSysMapper.selectByCondition(condition)).orElse(new ArrayList<>()).stream()
+                .filter(po -> {
+
+                        if (localFileSysPOList.stream().map(LocalFileSysPO::getPath).collect(Collectors.toList()).contains(po.getPath())) {
+                            return true;
+                        } else {
+                            try {
+                                FileUtils.forceDelete(new File(po.getPath()));
+                            } catch (IOException e) {
+                                e.printStackTrace();
+                            }
+                            return false;
+                        }
+
+                }).collect(Collectors.toList());
+
+        collect.forEach( po -> localFileSysMapper.delete(po));
+//        localFileSysMapper.deleteByCondition(condition);
         localFileSysMapper.insertList(localFileSysPOList);
         asyncServiceExecutor.execute(() -> {
             values.stream().forEach( item -> ZipFilesUtil.unZip(item.getOriginalFilename(), dest));
@@ -164,5 +181,103 @@ public class ClientController {
         localFileSysPO.setLevel((byte) 2);
         List<String> list = localFileSysMapper.select(localFileSysPO).stream().map(po -> po.getPath().substring(workPath.length())).collect(Collectors.toList());
         return new Result().setData(list);
+    }
+
+//    public String getPath(LocalFileSysPO po){
+//        String path = po.getPath();
+//        String subPath = path.substring(0, po.getPath().lastIndexOf("-") + 1);
+//        if(po.getType() == 0){
+//
+//            return subPath + "%";
+//        } else {
+//            String[] split = path.split(".");
+//            return subPath +
+//        }
+//    }
+
+    @PostMapping("uploadStdFile1")
+    @ApiIgnore
+    public String upload1(HttpServletRequest request, String dest){
+        MultipartHttpServletRequest multipartHttpServletRequest = null;
+        if(request instanceof MultipartHttpServletRequest){
+            multipartHttpServletRequest = (MultipartHttpServletRequest) request;
+        }
+        UploadFileUtil.multiUpload(request, dest);
+        Map<String, MultipartFile> fileMap = multipartHttpServletRequest.getFileMap();
+        Collection<MultipartFile> values = fileMap.values();
+        List<LocalFileSysPO> localFileSysPOList = new ArrayList<>();
+        LocalFileSysPO sysPO = new LocalFileSysPO();
+        sysPO.setPath(workPath + dest);
+        List<LocalFileSysPO> sysPOList = localFileSysMapper.select(sysPO);
+        values.stream().forEach(item -> {
+            String filename = item.getOriginalFilename();
+            String unzipDir = ZipFilesUtil.createUnzipDir(filename, dest);
+            boolean isDir = StringUtils.contains(filename, "$$") ? false : true;
+            LocalFileSysPO localFileSysPO = new LocalFileSysPO();
+            if(isDir){
+                localFileSysPO.setType((byte) 0);
+
+            } else {
+                localFileSysPO.setType((byte) 1);
+            }
+            localFileSysPO.setPath(unzipDir);
+            localFileSysPO.setFilename(unzipDir.substring(unzipDir.lastIndexOf("/") + 1));
+            localFileSysPO.setParentPath(workPath + dest);
+            localFileSysPO.setLevel((byte) 3);
+            localFileSysPO.setCreateTime(new Date());
+            localFileSysPO.setPid(sysPOList.get(0).getId());
+            localFileSysPO.setId(UUIDUtil.getUUID());
+            localFileSysPOList.add(localFileSysPO);
+
+        });
+
+        List<LocalFileSysPO> localFileSysPOListFromPreInsert = localFileSysPOList.stream().map(po -> {
+            LocalFileSysPO localFileSysPO = new LocalFileSysPO();
+            BeanUtils.copyProperties(po, localFileSysPO);
+            localFileSysPO.setFilename(po.getFilename().replaceAll(datePattern1, ""));
+            return localFileSysPO;
+        }).collect(Collectors.toList());
+        Condition condition = new Condition(LocalFileSysPO.class);
+        Example.Criteria criteria = condition.or();
+        localFileSysPOList.forEach(po -> criteria.orLike("path", po.getPath().substring(0, po.getPath().lastIndexOf("-") + 1) + "%"));
+        List<LocalFileSysPO> localFileSysPOFromDB = localFileSysMapper.selectByCondition(condition);
+        localFileSysPOFromDB.forEach(po -> {
+            String originalFilename = po.getFilename().replaceAll(datePattern1, "");
+            po.setFilename(originalFilename);
+        });
+        List<String> waitingDelete = new ArrayList<>();
+        for (int i = 0; i < localFileSysPOFromDB.size(); i++){
+            for (int j = 0; j < localFileSysPOListFromPreInsert.size(); j++){
+                if(StringUtils.equals(localFileSysPOFromDB.get(i).getFilename(), localFileSysPOListFromPreInsert.get(j).getFilename())){
+                    waitingDelete.add(localFileSysPOFromDB.get(i).getId());
+                    if(!StringUtils.equals(localFileSysPOFromDB.get(i).getPath(), localFileSysPOListFromPreInsert.get(j).getPath())){
+                        try {
+
+                            FileUtils.forceDelete(new File(localFileSysPOFromDB.get(i).getPath()));
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
+                    }
+                    break;
+                }
+            }
+        }
+        if(waitingDelete.size() > 0){
+            localFileSysMapper.deleteByIds(waitingDelete);
+        }
+
+
+        localFileSysMapper.insertList(localFileSysPOList);
+        asyncServiceExecutor.execute(() -> {
+            values.stream().forEach( item -> ZipFilesUtil.unZip(item.getOriginalFilename(), dest));
+            try {
+                FileUtils.cleanDirectory(new File(workPath + dest + "/temp"));
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        });
+        String remoteAddr = request.getRemoteAddr();
+        localFileSysPOList.forEach( item -> log.info("{}用户上传了文件{}", remoteAddr, item.getFilename()));
+        return "ok";
     }
 }
